# sha3: prevent state from escaping to heap