Two issues prevented the sha3/shake state from staying on the heap:

    1. xorIn and copyOut were virtuals (variables)
    2. State contains a pointer (buf) into itself

The solution is simple.

For (1) we set xorIn and copyOut at compile-time, which it already was,
but the compiler didn't notice.

For (2) we store the bounds of the original slice buf into storage
instead of the slice itself.

Actual performance benefit doesn't show up properly in this microbenchmark
as it is seen when there is a lot of pressure on the GC:

```
name                   old time/op   new time/op   delta
PermutationFunction-8    375ns ± 7%    375ns ± 9%    ~     (p=0.730 n=5+5)
Sha3_512_MTU-8          8.68µs ± 0%   8.10µs ± 3%  -6.74%  (p=0.008 n=5+5)
Sha3_384_MTU-8          6.27µs ± 1%   6.04µs ± 2%  -3.51%  (p=0.008 n=5+5)
Sha3_256_MTU-8          4.99µs ± 0%   4.72µs ± 4%  -5.56%  (p=0.008 n=5+5)
Sha3_224_MTU-8          4.78µs ± 0%   4.36µs ± 2%  -8.73%  (p=0.016 n=4+5)
Shake128_MTU-8          3.62µs ± 4%   3.59µs ± 3%    ~     (p=0.690 n=5+5)
Shake256_MTU-8          3.86µs ± 2%   3.86µs ± 3%    ~     (p=0.690 n=5+5)
Shake256_16x-8          53.9µs ± 2%   56.3µs ± 5%  +4.51%  (p=0.032 n=5+5)
Shake256_1MiB-8         2.96ms ± 2%   2.94ms ± 1%    ~     (p=0.421 n=5+5)
Sha3_512_1MiB-8         5.57ms ± 3%   5.45ms ± 1%    ~     (p=0.056 n=5+5)

name                   old speed     new speed     delta
PermutationFunction-8  533MB/s ± 7%  534MB/s ± 9%    ~     (p=0.841 n=5+5)
Sha3_512_MTU-8         155MB/s ± 0%  167MB/s ± 3%  +7.27%  (p=0.008 n=5+5)
Sha3_384_MTU-8         215MB/s ± 1%  223MB/s ± 2%  +3.66%  (p=0.008 n=5+5)
Sha3_256_MTU-8         270MB/s ± 0%  286MB/s ± 4%  +5.95%  (p=0.008 n=5+5)
Sha3_224_MTU-8         283MB/s ± 0%  310MB/s ± 2%  +9.59%  (p=0.016 n=4+5)
Shake128_MTU-8         373MB/s ± 4%  376MB/s ± 3%    ~     (p=0.690 n=5+5)
Shake256_MTU-8         350MB/s ± 2%  350MB/s ± 3%    ~     (p=0.690 n=5+5)
Shake256_16x-8         304MB/s ± 2%  291MB/s ± 5%  -4.25%  (p=0.032 n=5+5)
Shake256_1MiB-8        354MB/s ± 2%  357MB/s ± 1%    ~     (p=0.421 n=5+5)
Sha3_512_1MiB-8        188MB/s ± 3%  192MB/s ± 1%    ~     (p=0.056 n=5+5)
```

Closes #281