@bwesterb 2021-09-17
For Dilithium3 (v3.1) it makes a 13% difference:

```
name    old time/op  new time/op  delta
Sign-8   193µs ± 1%   166µs ± 7%  -13.96%  (p=0.008 n=5+5)
```